requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
pypdf>=4.0.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
python-dotenv>=1.0.0
